requirements.txt
PyMySQL==0.9.2
dnspython==1.15.0
PyYAML==3.13
